requirements.txt
streamlit>=1.30.0
folium>=0.15.0
streamlit-folium>=0.12.0
python-dotenv>=1.0.0
pandas>=2.0.0
